requirements.txt
discord.py
openai
flask
python-dotenv
